move writers to appropriate place
