+++ swiftwinrt/code_writers/writer_helpers.h
@@ -341,156 +341,8 @@ namespace swiftwinrt
 
     static std::vector<function_param> get_projected_params(attributed_type const& factory, function_def const& func);
     static write_scope_guard<writer> write_local_param_wrappers(writer& w, std::vector<function_param> const& params);
-    static void do_write_interface_abi(writer& w, typedef_base const& type, std::vector<function_def> const& methods)
-    {
-        auto factory_info = try_get_factory_info(w, type);
-        auto classType = try_get_exclusive_to(w, type);
 
-        if (classType && !can_write(w, classType))
-        {
-            return;
-        }
-
-        const bool isInitializer = factory_info.has_value() && (factory_info->activatable || factory_info->composable);
-        const bool composableFactory = factory_info.has_value() && factory_info->composable;
-
-        auto name = w.write_temp("%", type);
-        auto baseClass = (is_delegate(type) || !type.type().Flags().WindowsRuntime()) ? "IUnknown" : "IInspectable";
-        w.write("public class %: %.% {\n",
-            bind_type_abi(type),
-            w.support,
-            baseClass);
-
-        auto class_indent_guard = w.push_indent();
-
-        auto iid_format = "override public class var IID: %.IID { IID_% }\n\n";
-        w.write(iid_format, w.support, bind_type_mangled(type));
-
-        for (const auto& function : methods)
-        {
-            if (!can_write(w, function, true)) continue;
-            try
-            {
-                auto func_name = get_abi_name(function);
-                auto full_names = w.push_full_type_names(true);
-
-                auto returnStatement = isInitializer ?
-                    w.write_temp(" -> %", bind_type_abi(classType->default_interface)) :
-                    w.write_temp("%", bind<write_return_type_declaration>(function, write_type_params::swift));
-
-                std::vector<function_param> params = composableFactory ? get_projected_params(factory_info.value(), function) : function.params;
-                std::string written_params = w.write_temp("%", bind<write_function_params2>(params, write_type_params::swift));
-                if (composableFactory)
-                {
-                    if (params.size() > 0) written_params.append(", ");
-                    written_params.append(w.write_temp("_ baseInterface: UnsealedWinRTClassWrapper<%.Composable>?, _ innerInterface: inout %.IInspectable?", bind_bridge_name(*classType), w.support));
-                }
-
-                w.write("% func %(%) throws% {\n",
-                    is_exclusive(type) ? "public" : "open",
-                    func_name,
-                    written_params,
-                    returnStatement);
-                {
-                    auto function_indent_guard = w.push_indent();
-                    std::vector<writer::indent_guard> initialize_result_indent;
-                    if (function.return_type)
-                    {
-                        if (auto result = write_init_return_val_abi(w, function.return_type.value()))
-                        {
-                            initialize_result_indent.push_back(std::move(result.value()));
-                        }
-                    }
-
-                    {
-                        auto guard = write_local_param_wrappers(w, params);
-
-                        if (composableFactory)
-                        {
-                            w.write("let _baseInterface = baseInterface?.toIInspectableABI { $0 }\n");
-                            w.write("let (_innerInterface) = try ComPtrs.initialize { _innerInterfaceAbi in\n");
-                            guard.push_indent();
-                            guard.push("}\n");
-                            guard.push("innerInterface = %.IInspectable(_innerInterface!)\n", w.support);
-                        }
-                        w.write(R"(_ = try perform(as: %.self) { pThis in
-    try CHECKED(pThis.pointee.lpVtbl.pointee.%(%))
-}
-)",
-bind_type_mangled(type),
-func_name,
-bind<write_abi_args>(function));
-                    }
-
-                    for (auto&& guard : initialize_result_indent)
-                    {
-                        guard.end();
-                        w.write("}\n");
-                    }
-
-
-                    if (function.return_type && !isInitializer)
-                    {
-                        w.write("%\n", bind<write_consume_return_statement>(function));
-                    }
-                    else if (isInitializer)
-                    {
-                        w.write("return %(%!)\n", bind_type_abi(classType->default_interface), function.return_type->name);
-                    }
-                }
-                w.write("}\n\n");
-            }
-            catch (std::exception const& e)
-            {
-                throw_invalid(e.what(),
-                    "\n method: ", get_name(function),
-                    "\n type: ", get_full_type_name(type),
-                    "\n database: ", type.type().get_database().path());
-            }
-        }
-
-        class_indent_guard.end();
-        w.write("}\n\n");
-    }
-
-    static void write_implementable_interface(writer& w, interface_type const& type);
-
-
-    static void write_class_func_body(writer& w, function_def const& function, interface_info const& iface, bool is_noexcept);
     static void write_comma_param_names(writer& w, std::vector<function_param> const& params);
-    template <typename T>
-    static void write_delegate_extension(writer& w, T const& inst, function_def const& invoke_method)
-    {
-        auto delegate_abi_name = w.write_temp("%", bind_type_mangled(inst));
-
-        constexpr bool is_generic = std::is_same_v<T, generic_inst>;
-        w.write("@_spi(WinRTInternal)\n");
-        w.write(R"(% extension WinRTDelegateBridge where CABI == % {
-    static func makeAbi() -> CABI {
-        let vtblPtr = withUnsafeMutablePointer(to: &%.%VTable) { $0 }
-        return .init(lpVtbl:vtblPtr)
-    }
-}
-
-)", is_generic ? "internal" : "public",
-    delegate_abi_name,
-    is_generic ? w.swift_module : abi_namespace(w.type_namespace),
-    is_generic ? delegate_abi_name : w.write_temp("%", inst));
-    }
-
-    static void write_vtable(writer& w, interface_type const& type);
-    static void write_vtable(writer& w, delegate_type const& type);
-
-    template <typename T>
-    static void write_delegate_wrapper(writer& w, T const& type)
-    {
-        auto impl_name = w.write_temp("%", bind_bridge_fullname(type));
-        auto wrapper_name = w.write_temp("%", bind_wrapper_name(type));
-        auto format = R"(
-typealias % = InterfaceWrapperBase<%>
-)";
-        w.write(format, wrapper_name, impl_name);
-    }
 
     static void write_convert_vtable_params(writer& w, function_def const& signature)
     {
@@ -601,77 +453,6 @@ typealias % = InterfaceWrapperBase<%>
         }
     }
 
-    static void write_implementable_interface(writer& w, interface_type const& type)
-    {
-        write_vtable(w, type);
-
-        // Define a struct that matches a C++ object with a vtable. As background:
-        //
-        // in C++ this looks like:
-        //
-        //   class CMyObject : public IMyInterface {
-        //   {
-        //     HRESULT Foo(int number);
-        //   }
-        //
-        // in C it looks
-        //
-        //   struct __x_ABI_MyObjectVTable {
-        //     ... AddRef, Release, QI, ...
-        //     HRESULT (STDMETHODCALLTYPE * Foo)(__x_ABI_IMyInterface* pThis, int number)
-        //   }
-        //
-        //   struct __x_ABI_IMyInterface {
-        //     const __x_ABI_MyObjectVTable* lpVtbl;
-        //   }
-        //
-        // so in Swift we're using the pattern:
-        //
-        //   protocol IMyInterface
-        //   {
-        //      func Foo(number: Int32)
-        //   }
-        //
-        //   private var myInterfaceVTable: __x_ABI_MyObjectVTable {
-        //     Foo: {
-        //       // In C, 'pThis' is always the first param
-        //       guard let instance = MyInterfaceWrapper.try_unwrap_raw($0)?.takeUnretainedValue().swiftObj else {
-        //         return E_INVALIDARG
-        //       }
-        //       let number = $1
-        //       instance.Foo(number)
-        //     }
-        //   }
-        //   ...
-        //    class MyInterfaceWrapper : WinRTWrapperBase<__x_ABI_IMyInterface, IMyInterface> {
-        //       init(impl: IMyInterface) {
-        //         let abi = withUnsafeMutablePointer(to: &myInterfaceVTable) {
-        //           __x_ABI_IMyInterface(lpVtbl: $0)
-        //         }
-        //       super.init(abi, impl)
-        //    }
-        // Where the WinRTWrapperBase defines the behavior for all objects which need to wrap a
-        // swift object and pass it down to WinRT:
-        // open class WinRTWrapperBase<CInterface, Prototype> {
-        //    public struct ComObject {
-        //      public var comInterface : CInterface
-        //      public var wrapper : Unmanaged<WinRTWrapperBase>?
-        //    }
-        //    public var instance : ComObject
-        //    public var swiftObj : Prototype
-        //    ...
-        // }
-
-        w.write(R"(
-public typealias % = InterfaceWrapperBase<%>
-)",
-bind_wrapper_name(type),
-bind_bridge_fullname(type));
-    }
-
-    static void write_class_impl_func(writer& w, function_def const& method, interface_info const& iface, typedef_base const& type_definition);
-    static void write_class_impl_property(writer& w, property_def const& prop, interface_info const& iface, typedef_base const& type_definition);
-    static void write_class_impl_event(writer& w, event_def const& event, interface_info const& iface, typedef_base const& type_definition);
     static void write_property_value_impl(writer& w)
     {
         auto winrtInterfaceConformance = w.write_temp(R"(
@@ -866,71 +647,6 @@ bind_bridge_fullname(type));
 )", modifier_for(type_definition, info), bufferType,  w.support, type.swift_type_name(), get_swift_name(info));
     }
 
-    static void write_interface_impl_members(writer& w, interface_info const& info, typedef_base const& type_definition)
-    {
-        w.add_depends(*info.type);
-        bool is_class = swiftwinrt::is_class(&type_definition);
-
-        if (!info.is_default || (!is_class && info.base))
-        {
-            auto swiftAbi = w.write_temp("%.%", abi_namespace(info.type->swift_logical_namespace()), info.type->swift_type_name());
-            if (is_generic_inst(info.type))
-            {
-                auto guard{ w.push_generic_params(info) };
-                swiftAbi = w.write_temp("%", bind_type_abi(info.type));
-            }
-            w.write("private lazy var %: %! = getInterfaceForCaching()\n",
-                get_swift_name(info),
-                swiftAbi);
-        }
-
-        if (auto iface = dynamic_cast<const interface_type*>(info.type))
-        {
-            for (const auto& method : iface->functions)
-            {
-                write_class_impl_func(w, method, info, type_definition);
-            }
-
-            for (const auto& prop : iface->properties)
-            {
-                write_class_impl_property(w, prop, info, type_definition);
-            }
-
-            for (const auto& event : iface->events)
-            {
-                write_class_impl_event(w, event, info, type_definition);
-            }
-        }
-        else if (auto gti = dynamic_cast<const generic_inst*>(info.type))
-        {
-            for (const auto& method : gti->functions)
-            {
-                write_class_impl_func(w, method, info, type_definition);
-            }
-
-            for (const auto& prop : gti->properties)
-            {
-                write_class_impl_property(w, prop, info, type_definition);
-            }
-
-            for (const auto& event : gti->events)
-            {
-                write_class_impl_event(w, event, info, type_definition);
-            }
-        }
-        else if (auto systemType = dynamic_cast<const system_type*>(info.type))
-        {
-            if (systemType->swift_type_name() == "IBufferByteAccess" || systemType->swift_type_name() == "IMemoryBufferByteAccess")
-            {
-                write_bufferbyteaccess(w, info, *systemType, type_definition);
-            }
-        }
-        else
-        {
-            assert(!"Unexpected interface type.");
-        }
-    }
-
     static bool skip_write_from_abi(writer& w, metadata_type const& type)
     {
         if (auto interfaceType = dynamic_cast<const interface_type*>(&type))
@@ -942,22 +658,6 @@ bind_bridge_fullname(type));
             return classType->default_interface == nullptr;
         }
         return true;
-    }
-
-    static void write_comma_param_types(writer& w, std::vector<function_param> const& params);
-    static void write_delegate_return_type(writer& w, function_def const& sig);
-    static void write_query_interface_case(writer& w, interface_info const& iface);
-
-    static void write_delegate_return_type(writer& w, function_def const& sig)
-    {
-        if (sig.return_type)
-        {
-            write_type(w, *sig.return_type->type, write_type_params::swift);
-        }
-        else
-        {
-            w.write("()");
-        }
     }
 
     static void write_comma_param_names(writer& w, std::vector<function_param> const& params)
@@ -986,53 +686,6 @@ bind_bridge_fullname(type));
             }
         }
     }
-
-
-    template <typename T>
-    static void do_write_delegate_implementation(writer& w, T const& type, function_def const& invoke_method)
-    {
-        auto format = R"(% class % : WinRTDelegateBridge {
-    % typealias Handler = %
-    % typealias CABI = %
-    % typealias SwiftABI = %.%
-
-    % static func from(abi: ComPtr<CABI>?) -> Handler? {
-        guard let abi = abi else { return nil }
-        let _default = SwiftABI(abi)
-        let handler: Handler = { (%) in
-%        }
-        return handler
-    }
-}
-)";
-        auto data = w.write_temp("%", bind<write_comma_param_types>(invoke_method.params));
-        auto return_type = w.write_temp("%", bind<write_delegate_return_type>(invoke_method));
-        constexpr bool is_generic = std::is_same_v<T, generic_inst>;
-        auto access_level = is_generic ? "internal" : "public";
-        auto handlerType = w.write_temp("%", bind<write_swift_type_identifier>(type));
-        auto abi_guard = w.push_abi_types(is_generic);
-        w.write(format,
-            access_level,
-            bind_bridge_name(type),
-            access_level,
-            handlerType,
-            access_level,
-            bind_type_mangled(type),
-            access_level,
-            is_generic ? w.swift_module : abi_namespace(w.type_namespace),
-            bind_type_abi(type),
-            access_level,
-            bind<write_comma_param_names>(invoke_method.params),
-            bind([&](writer& w) {
-                interface_info delegate{ &type };
-                delegate.is_default = true; // so that the _default name is used
-                auto indent_guard{ w.push_indent({3}) };
-                write_class_func_body(w, invoke_method, delegate, false);
-                }));
-    }
-
-
-    void write_interface_bridge(writer& w, metadata_type const& type);
 
     static void write_param_names(writer& w, std::vector<function_param> const& params, std::string_view format)
     {
@@ -1387,169 +1040,6 @@ bind_bridge_fullname(type));
         return false;
     }
 
-    static void write_factory_constructors(writer& w, attributed_type const& factory, class_type const& type, metadata_type const& default_interface)
-    {
-        if (auto factoryIface = dynamic_cast<const interface_type*>(factory.type))
-        {
-            interface_info factory_info{ factoryIface };
-            auto swift_name = get_swift_name(factory_info);
-            w.write("private static let %: %.% = try! RoGetActivationFactory(\"%\")\n",
-                swift_name, abi_namespace(factoryIface), factory.type, get_full_type_name(type));
-            for (const auto& method : factoryIface->functions)
-            {
-                if (!can_write(w, method)) continue;
-
-                auto baseHasMatchingConstructor = base_has_matching_constructor(type, factory, method);
-                w.write("%public init(%) {\n",
-                    baseHasMatchingConstructor ? "override " : "",
-                    bind<write_function_params>(method, write_type_params::swift_allow_implicit_unwrap));
-                {
-                    auto indent = w.push_indent();
-                    write_factory_body(w, method, factory_info, type, default_interface);
-                }
-                w.write("}\n\n");
-            }
-        }
-        else
-        {
-            auto base_class = type.base_class;
-
-            w.write("private static let _defaultFactory: %.IActivationFactory = try! RoGetActivationFactory(\"%\")\n",
-                w.support, get_full_type_name(type));
-            w.write("%public init() {\n", has_default_constructor(base_class) ? "override " : "");
-            {
-                auto indent = w.push_indent();
-                auto activateInstance = "try! Self._defaultFactory.ActivateInstance()";
-                if (base_class)
-                {
-                    w.write("super.init(fromAbi: %)\n", activateInstance);
-                }
-                else
-                {
-                    w.write("super.init(%)\n", activateInstance);
-                }
-            }
-            w.write("}\n\n");
-        }
-    }
-
-    // every composable type needs a special composing constructor to help satisfy swift type initializer
-    // requirements. this constructor is marked as SPI so that it doesn't show up to normal developers
-    // when they are building.
-    static void write_internal_composable_constructor(writer& w, class_type const& type)
-    {
-        if (type.base_class)
-        {
-            w.write(R"(^@_spi(WinRTInternal)
-override public init<Composable: ComposableImpl>(
-    composing: Composable.Type,
-    _ createCallback: (UnsealedWinRTClassWrapper<Composable>?, inout %.IInspectable?) -> Composable.Default.SwiftABI)
-{
-    super.init(composing: composing, createCallback)
-}
-)", w.support);
-        }
-        else
-        {
-            w.write(R"(^@_spi(WinRTInternal)
-public init<Composable: ComposableImpl>(
-    composing: Composable.Type,
-    _ createCallback: (UnsealedWinRTClassWrapper<Composable>?, inout %.IInspectable?) -> Composable.Default.SwiftABI)
-{
-    super.init()
-    MakeComposed(composing: composing, (self as! Composable.Class), createCallback)
-}
-)", w.support);
-        }
-    }
-
-    static void write_composable_constructor(writer& w, attributed_type const& factory, class_type const& type)
-    {
-        if (auto factoryIface = dynamic_cast<const interface_type*>(factory.type))
-        {
-            w.write("private static var _% : %.% =  try! RoGetActivationFactory(\"%\")\n\n",
-                    factory.type,
-                    abi_namespace(factoryIface),
-                    factory.type,
-                    get_full_type_name(type));
-
-            interface_info factory_info{ factoryIface };
-            for (const auto& method : factoryIface->functions)
-            {
-                if (!can_write(w, method)) continue;
-
-                auto baseHasMatchingConstructor = base_has_matching_constructor(type, factory, method);
-
-                std::vector<function_param> params = get_projected_params(factory, method);
-
-                w.write("%public init(%) {\n", baseHasMatchingConstructor ? "override " : "", bind<write_function_params2>(params, write_type_params::swift_allow_implicit_unwrap));
-                {
-                    auto indent = w.push_indent();
-                    std::string_view func_name = get_abi_name(method);
-
-                    auto return_name = method.return_type.value().name;
-                    {
-                        if (type.base_class)
-                        {
-                            w.write("super.init(composing: %.Composable.self) { baseInterface, innerInterface in \n", bind_bridge_name(type));
-                        }
-                        else
-                        {
-                            w.write("super.init()\n");
-                            w.write("MakeComposed(composing: %.Composable.self, self) { baseInterface, innerInterface in \n", bind_bridge_name(type));
-                        }
-                        w.write("    try! Self.%.%(%)\n",
-                            get_swift_name(factory_info),
-                            func_name,
-                            bind<write_implementation_args>(method));
-                        w.write("}\n");
-                    }
-                }
-                w.write("}\n\n");
-            }
-        }
-    }
-
-    static void write_composable_impl(writer& w, class_type const& parent, metadata_type const& overrides, bool compose);
-
-    static void write_default_constructor_declarations(writer& w, class_type const& type, metadata_type const& default_interface)
-    {
-        auto base_class = type.base_class;
-        w.write("@_spi(WinRTInternal)\n");
-        w.write("%public init(fromAbi: %.IInspectable) {\n",
-            base_class ? "override " : "",
-            w.support);
-        {
-            auto indent = w.push_indent();
-            if (base_class)
-            {
-                w.write("super.init(fromAbi: fromAbi)\n");
-            }
-            else
-            {
-                w.write("super.init(fromAbi)\n");
-            }
-        }
-        w.write("}\n\n");
-
-        if (type.is_composable())
-        {
-            write_internal_composable_constructor(w, type);
-        }
-    }
-
-    static void write_class_func_body(writer& w, function_def const& function, interface_info const& iface, bool is_noexcept)
-    {
-        std::string_view func_name = get_abi_name(function);
-        auto impl = get_swift_name(iface);
-        auto try_flavor = is_noexcept ? "try!" : "try";
-        w.write("% %.%(%)\n",
-            try_flavor,
-            impl,
-            func_name,
-            bind<write_implementation_args>(function));
-    }
-
     static std::string modifier_for(typedef_base const& type_definition, interface_info const& iface, member_type member)
     {
         std::string modifier;
@@ -1596,264 +1086,6 @@ public init<Composable: ComposableImpl>(
             }
         }
         return modifier;
-    }
-
-
-    static void write_class_impl_property(writer& w, property_def const& prop, interface_info const& iface, typedef_base const& type_definition)
-    {
-        if (!can_write(w, prop)) return;
-
-        write_documentation_comment(w, type_definition, prop.def.Name());
-
-        auto impl = get_swift_name(iface);
-
-        if (prop.getter)
-        {
-            auto format = prop.is_array() ? "[%]" : "%";
-            auto propertyType = w.write_temp(format, bind<write_type>(*prop.getter->return_type->type, swift_write_type_params_for(*iface.type, prop.is_array())));
-            w.write("%var % : % {\n",
-                modifier_for(type_definition, iface),
-                get_swift_name(prop),
-                propertyType);
-
-            w.write("    get { try! %.%() }\n",
-                impl,
-                get_swift_name(prop.getter.value()));
-
-            // TODO: https://linear.app/the-browser-company/issue/WIN-82/support-setters-not-defined-in-same-api-contract-as-getters
-            // right now require that both getter and setter are defined in the same version
-            if (prop.setter)
-            {
-                w.write("    set { try! %.%(newValue) }\n", impl, get_swift_name(prop.setter.value()));
-
-            }
-            w.write("}\n\n");
-        }
-    }
-
-    static void write_class_impl_func(writer& w, function_def const& function, interface_info const& iface, typedef_base const& type_definition)
-    {
-        if (function.def.SpecialName() || !can_write(w, function))
-        {
-            // don't write methods which are really properties
-            return;
-        }
-
-        write_documentation_comment(w, type_definition, function.def.Name());
-        auto is_no_except = is_noexcept(*iface.type, function);
-        auto type_params = swift_write_type_params_for(*iface.type);
-        auto maybe_throws = is_no_except ? "" : " throws";
-        w.write("%func %(%)%% {\n",
-            modifier_for(type_definition, iface),
-            get_swift_name(function),
-            bind<write_function_params>(function, type_params),
-            maybe_throws,
-            bind<write_return_type_declaration>(function, type_params));
-        {
-            auto indent = w.push_indent();
-            write_class_func_body(w, function, iface, is_no_except);
-        }
-        w.write("}\n\n");
-    }
-
-    static void write_class_impl_event(writer& w, event_def const& def, interface_info const& iface, typedef_base const& type_definition)
-    {
-        write_documentation_comment(w, type_definition, def.def.Name());
-
-        auto event = def.def;
-        auto format = R"(%var % : Event<%> = {
-  .init(
-    add: { [weak self] in
-      guard let this = self?.% else { return .init() }
-      return try! this.add_%($0)
-    },
-    remove: { [weak self] in
-     try? self?.%.remove_%($0)
-   }
-  )
-}()
-
-)";
-
-        auto static_format = R"(%var % : Event<%> = {
-  .init(
-    add: { try! %.add_%($0) },
-    remove: { try? %.remove_%($0) }
-  )
-}()
-
-)";
-        auto type = find_type(event.EventType());
-        writer::generic_param_guard guard{};
-        function_def delegate_method{};
-        if (auto delegateType = dynamic_cast<const delegate_type*>(def.type))
-        {
-            delegate_method = delegateType->functions[0];
-        }
-        else if (auto genericInst = dynamic_cast<const generic_inst*>(def.type))
-        {
-            delegate_method = genericInst->functions[0];
-            guard = w.push_generic_params(*genericInst);
-        }
-
-        auto modifier = modifier_for(type_definition, iface, member_type::event);
-        if (!iface.attributed)
-        {
-            modifier.append("lazy ");
-        }
-        assert(delegate_method.def);
-        w.write(iface.attributed ? static_format : format,
-            modifier, // % var
-            get_swift_name(event), // var %
-            def.type, // Event<%>
-            get_swift_name(iface), // weak this = %
-            def.def.Name(), // add_&Impl
-            get_swift_name(iface), // weak this = %
-            def.def.Name() // remove_&Impl
-        );
-    }
-
-    static void write_statics_body(writer& w, function_def const& method, metadata_type const& statics)
-    {
-        std::string_view func_name = get_abi_name(method);
-
-        if (method.return_type)
-        {
-            w.write("return ");
-        }
-        w.write("try _%.%(%)\n",
-            statics.swift_type_name(),
-            func_name,
-            bind<write_implementation_args>(method));
-    }
-
-    static void write_static_members(writer& w, attributed_type const& statics, class_type const& type)
-    {
-        if (auto ifaceType = dynamic_cast<const interface_type*>(statics.type))
-        {
-            interface_info static_info{ statics.type };
-            static_info.attributed = true;
-
-            auto impl_name = get_swift_name(static_info);
-            w.write("private static let %: %.% = try! RoGetActivationFactory(\"%\")\n",
-                impl_name,
-                abi_namespace(statics.type),
-                statics.type->swift_type_name(),
-                get_full_type_name(type));
-
-            for (const auto& method : ifaceType->functions)
-            {
-                if (!can_write(w, method))
-                {
-                    continue;
-                }
-
-                write_documentation_comment(w, type, method.def.Name());
-                w.write("%func %(%) throws% {\n",
-                    modifier_for(type, statics, method),
-                    get_swift_name(method),
-                    bind<write_function_params>(method, write_type_params::swift_allow_implicit_unwrap),
-                    bind<write_return_type_declaration>(method, write_type_params::swift_allow_implicit_unwrap));
-                {
-                    auto indent = w.push_indent();
-                    write_statics_body(w, method, *statics.type);
-                }
-                w.write("}\n\n");
-            }
-
-            for (const auto& static_prop : ifaceType->properties)
-            {
-                write_class_impl_property(w, static_prop, static_info, type);
-            }
-
-            for (const auto& event : ifaceType->events)
-            {
-                write_class_impl_event(w, event, static_info, type);
-            }
-        }
-    }
-
-    static void write_composable_impl(writer& w, class_type const& parent, metadata_type const& overrides, bool compose)
-    {
-        auto default_interface = parent.default_interface;
-        if (!default_interface)
-        {
-            throw_invalid("Could not find default interface for %s\n", parent.swift_type_name().data());
-        }
-
-        bool use_iinspectable_vtable = type_name(overrides) == type_name(*default_interface);
-
-        auto format = R"(public enum % : ComposableImpl {
-    public typealias CABI = %
-    public typealias SwiftABI = %
-    public typealias Class = %
-    public typealias SwiftProjection = WinRTClassWeakReference<Class>
-    public enum Default : AbiInterface {
-        public typealias CABI = %
-        public typealias SwiftABI = %
-    }
-}
-)";
-
-        auto composableName = w.write_temp("%", bind_type_abi(overrides));
-        // If we're composing a type without any overrides, then we'll just create an IInspectable vtable which wraps
-        // this object and provides facilities for reference counting and getting the class name of the swift object.
-        w.write(format,
-            composableName,
-            bind([&](writer& w) {
-                if (use_iinspectable_vtable)
-                {
-                    write_type_mangled(w, ElementType::Object);
-                }
-                else
-                {
-                    write_type_mangled(w, overrides);
-                }}),
-            bind([&](writer& w) {
-                if (use_iinspectable_vtable)
-                {
-                    w.write("%.IInspectable", w.support);
-                }
-                else
-                {
-                    w.write("%.%", abi_namespace(overrides), composableName);
-                }}),
-            parent,
-            bind_type_mangled(default_interface),
-            bind([&](writer& w) {
-                if (is_generic_inst(overrides))
-                {
-                    w.write("%.%", w.swift_module, composableName);
-                }
-                else
-                {
-                    w.write("%.%", abi_namespace(parent), default_interface);
-                }
-            }));
-
-        if (compose)
-        {
-            w.write("@_spi(WinRTInternal)\n");
-            w.write("public typealias Composable = %\n", composableName);
-        }
-    }
-
-    // write the default implementation for makeAbi. this way we don't need to expose the internal implementation
-    // details of the vtable to external modules.
-
-    static void write_query_interface_case(writer& w, interface_info const& iface)
-    {
-        w.write("case %.IID:\n", bind_wrapper_fullname(iface.type));
-        w.write("    let wrapper = %(self)\n", bind_wrapper_fullname(iface.type));
-        w.write("    return wrapper!.queryInterface(iid)\n");
-    }
-
-    static void write_iunknown_methods(writer& w, metadata_type const& type)
-    {
-        auto wrapper_name = w.write_temp("%", bind_wrapper_name(type));
-        w.write("QueryInterface: { %.queryInterface($0, $1, $2) },\n", wrapper_name);
-        w.write("AddRef: { %.addRef($0) },\n", wrapper_name);
-        w.write("Release: { %.release($0) },\n", wrapper_name);
     }
 
     template <typename T>
@@ -1922,7 +1154,7 @@ public init<Composable: ComposableImpl>(
 
     // assigns return or out parameters in vtable methods
     template<typename T>
-    static void do_write_abi_val_assignment(writer& w, T const& return_type, std::string_view return_param_name)
+    static void write_abi_val_assignment(writer& w, T const& return_type, std::string_view return_param_name)
     {
         auto type = return_type.type;
         auto param_name = get_swift_member_name(return_type.name);
@@ -2000,7 +1232,7 @@ public init<Composable: ComposableImpl>(
             if (param.out())
             {
                 auto return_param_name = "$" + std::to_string(param_number);
-                do_write_abi_val_assignment(w, param, return_param_name);
+                write_abi_val_assignment(w, param, return_param_name);
             }
 
             param_number++;
@@ -2014,7 +1246,7 @@ public init<Composable: ComposableImpl>(
                 param_number++;
             }
             auto return_param_name = "$" + std::to_string(param_number);
-            do_write_abi_val_assignment(w, signature.return_type.value(), return_param_name);
+            write_abi_val_assignment(w, signature.return_type.value(), return_param_name);
         }
     }
 
@@ -2124,7 +1356,7 @@ public init<Composable: ComposableImpl>(
     }
 
     template <typename T>
-    static void do_write_vtable(writer& w, T const& type, std::vector<named_interface_info> interfaces)
+    static void write_vtable_body(writer& w, T const& type, std::vector<named_interface_info> interfaces)
     {
         constexpr bool isInterface = std::is_same_v<T, interface_type>;
         constexpr bool isDelegate = std::is_same_v<T, delegate_type>;
@@ -2160,52 +1392,5 @@ public init<Composable: ComposableImpl>(
 )
 )");
 
-    }
-    static void write_vtable(writer& w, interface_type const& type)
-    {
-        do_write_vtable(w, type, type.required_interfaces);
-    }
-
-    static void write_vtable(writer& w, delegate_type const& type)
-    {
-        do_write_vtable(w, type, {});
-    }
-
-    static void write_overrides_vtable(writer& w, class_type const& type, interface_info const& overrides, std::vector<named_interface_info> const& other_interfaces)
-    {
-        w.write("internal typealias % = UnsealedWinRTClassWrapper<%.%>\n",
-            bind_wrapper_name(overrides.type),
-            bind_bridge_fullname(type),
-            overrides.type
-        );
-        w.write("internal static var %VTable: %Vtbl = .init(\n",
-            overrides.type,
-            bind_type_mangled(overrides.type));
-
-        {
-            auto indent = w.push_indent();
-            write_iunknown_methods(w, *overrides.type);
-            write_iinspectable_methods(w, overrides.type, other_interfaces, true);
-
-            separator s{ w, ",\n\n" };
-            s(); // get first separator out of the way for no-op
-
-            if (auto iface = dynamic_cast<const interface_type*>(overrides.type))
-            {
-                for (const auto& method : iface->functions)
-                {
-                    if (method.def.Name() != ".ctor")
-                    {
-                        s();
-                        write_vtable_method(w, method, *iface);
-                    }
-                }
-            }
-
-        }
-
-        w.write(R"(
-)
-)");
     }
 }
