+++ tests/test_component/Sources/test_component/Windows.Foundation+DeferralCompletedHandler.swift
@@ -56,7 +56,6 @@ extension __ABI_Windows_Foundation {
         }
     )
 }
-@_spi(WinRTInternal)
 public extension WinRTDelegateBridge where CABI == __x_ABI_CWindows_CFoundation_CIDeferralCompletedHandler {
     static func makeAbi() -> CABI {
         let vtblPtr = withUnsafeMutablePointer(to: &__ABI_Windows_Foundation.DeferralCompletedHandlerVTable) { $0 }
