+++ tests/test_component/Sources/test_component/test_component+ObjectHandler.swift
@@ -59,7 +59,6 @@ extension __ABI_test_component {
         }
     )
 }
-@_spi(WinRTInternal)
 public extension WinRTDelegateBridge where CABI == __x_ABI_Ctest__component_CIObjectHandler {
     static func makeAbi() -> CABI {
         let vtblPtr = withUnsafeMutablePointer(to: &__ABI_test_component.ObjectHandlerVTable) { $0 }
