+++ tests/test_component/Sources/test_component/test_component.Delegates+InObjectDelegate.swift
@@ -59,7 +59,6 @@ extension __ABI_test_component_Delegates {
         }
     )
 }
-@_spi(WinRTInternal)
 public extension WinRTDelegateBridge where CABI == __x_ABI_Ctest__component_CDelegates_CIInObjectDelegate {
     static func makeAbi() -> CABI {
         let vtblPtr = withUnsafeMutablePointer(to: &__ABI_test_component_Delegates.InObjectDelegateVTable) { $0 }
