+++ tests/test_component/Sources/test_component/test_component.Delegates+SignalDelegate.swift
@@ -56,7 +56,6 @@ extension __ABI_test_component_Delegates {
         }
     )
 }
-@_spi(WinRTInternal)
 public extension WinRTDelegateBridge where CABI == __x_ABI_Ctest__component_CDelegates_CISignalDelegate {
     static func makeAbi() -> CABI {
         let vtblPtr = withUnsafeMutablePointer(to: &__ABI_test_component_Delegates.SignalDelegateVTable) { $0 }
